+++ packages/jest-cli/src/__tests__/watch.test.js
@@ -65,7 +65,7 @@ jest.doMock(
     class WatchPlugin1 {
       getUsageInfo() {
         return {
-          key: 's'.codePointAt(0),
+          key: 's',
           prompt: 'do nothing',
         };
       }
@@ -79,7 +79,7 @@ jest.doMock(
     class WatchPlugin2 {
       getUsageInfo() {
         return {
-          key: 'u'.codePointAt(0),
+          key: 'u',
           prompt: 'do something else',
         };
       }
@@ -90,7 +90,6 @@ jest.doMock(
 const watch = require('../watch').default;
 
 const nextTick = () => new Promise(res => process.nextTick(res));
-const toHex = char => Number(char.charCodeAt(0)).toString(16);
 
 afterEach(runJestMock.mockReset);
 
@@ -329,7 +328,7 @@ describe('Watch mode flows', () => {
           }
           getUsageInfo() {
             return {
-              key: 'p'.codePointAt(0),
+              key: 'p',
               prompt: 'custom "P" plugin',
             };
           }
@@ -352,7 +351,7 @@ describe('Watch mode flows', () => {
 
     expect(pipe.write.mock.calls.reverse()[0]).toMatchSnapshot();
 
-    stdin.emit(toHex('p'));
+    stdin.emit('p');
     await nextTick();
 
     expect(run).toHaveBeenCalled();
@@ -405,7 +404,7 @@ describe('Watch mode flows', () => {
           }
           getUsageInfo() {
             return {
-              key: 's'.codePointAt(0),
+              key: 's',
               prompt: 'do nothing',
             };
           }
@@ -424,7 +423,7 @@ describe('Watch mode flows', () => {
       stdin,
     );
 
-    stdin.emit(Number('s'.charCodeAt(0)).toString(16));
+    stdin.emit('s');
 
     await nextTick();
 
@@ -445,7 +444,7 @@ describe('Watch mode flows', () => {
           onKey() {}
           getUsageInfo() {
             return {
-              key: 's'.codePointAt(0),
+              key: 's',
               prompt: 'do nothing',
             };
           }
@@ -465,7 +464,7 @@ describe('Watch mode flows', () => {
           onKey() {}
           getUsageInfo() {
             return {
-              key: 'z'.codePointAt(0),
+              key: 'z',
               prompt: 'also do nothing',
             };
           }
@@ -484,14 +483,14 @@ describe('Watch mode flows', () => {
       stdin,
     );
 
-    stdin.emit(Number('s'.charCodeAt(0)).toString(16));
+    stdin.emit('s');
     await nextTick();
     expect(run).toHaveBeenCalled();
-    stdin.emit(Number('z'.charCodeAt(0)).toString(16));
+    stdin.emit('z');
     await nextTick();
     expect(showPrompt2).not.toHaveBeenCalled();
     await resolveShowPrompt();
-    stdin.emit(Number('z'.charCodeAt(0)).toString(16));
+    stdin.emit('z');
     expect(showPrompt2).toHaveBeenCalled();
   });
 
@@ -537,7 +536,7 @@ describe('Watch mode flows', () => {
     runJestMock.mockReset();
 
     stdin.emit(KEYS.T);
-    ['t', 'e', 's', 't'].map(toHex).forEach(key => stdin.emit(key));
+    ['t', 'e', 's', 't'].forEach(key => stdin.emit(key));
     stdin.emit(KEYS.ENTER);
     await nextTick();
 
@@ -556,7 +555,7 @@ describe('Watch mode flows', () => {
     runJestMock.mockReset();
 
     stdin.emit(KEYS.P);
-    ['f', 'i', 'l', 'e'].map(toHex).forEach(key => stdin.emit(key));
+    ['f', 'i', 'l', 'e'].forEach(key => stdin.emit(key));
     stdin.emit(KEYS.ENTER);
     await nextTick();
 
@@ -575,12 +574,12 @@ describe('Watch mode flows', () => {
     runJestMock.mockReset();
 
     stdin.emit(KEYS.P);
-    ['f', 'i', 'l', 'e'].map(toHex).forEach(key => stdin.emit(key));
+    ['f', 'i', 'l', 'e'].forEach(key => stdin.emit(key));
     stdin.emit(KEYS.ENTER);
     await nextTick();
 
     stdin.emit(KEYS.T);
-    ['t', 'e', 's', 't'].map(toHex).forEach(key => stdin.emit(key));
+    ['t', 'e', 's', 't'].forEach(key => stdin.emit(key));
     stdin.emit(KEYS.ENTER);
     await nextTick();